+++ Panther/src/symbol_proc/SymbolProc.cpp
@@ -88,26 +88,6 @@ namespace pcit::panther{
 	}
 
 
-
-	auto SymbolProc::waitOnPIRLowerIfNeeded(ID id, Context& context, ID self_id) -> WaitOnResult {
-		const auto wait_on_lock = std::scoped_lock(wait_on_if_needed_lock);
-
-		if(this->isPIRLowerDone()){ return WaitOnResult::NOT_NEEDED; }
-		if(this->errored){ return WaitOnResult::WAS_ERRORED; }
-
-		SymbolProc& waiting_symbol = context.symbol_proc_manager.getSymbolProc(id);
-
-		const auto lock = std::scoped_lock(this->pir_lower_waited_on_lock, waiting_symbol.waiting_for_lock);
-
-		if(this->pir_lower_done){ return WaitOnResult::NOT_NEEDED; }
-
-		this->pir_lower_waited_on_by.emplace_back(id);
-		waiting_symbol.waiting_for.emplace_back(self_id);
-
-		return WaitOnResult::WAITING;
-	}
-
-
 	auto SymbolProc::waitOnPIRReadyIfNeeded(ID id, Context& context, ID self_id) -> WaitOnResult {
 		const auto wait_on_lock = std::scoped_lock(wait_on_if_needed_lock);